Initial tkinter test changes
