--- src/snooker-ball-tracker/config/__init__.py
@@ -1,8 +1,0 @@
-from os.path import dirname, basename, isfile
-import glob
-
-modules = glob.glob(dirname(__file__)+"/*.py")
-__all__ = [ basename(f)[:-3] for f in modules if isfile(f) and not f.endswith('__init__.py')]
-
-def get_setting_modules():
-    return __all__
