+++ src/snooker_ball_tracker/__main__.py
@@ -1,0 +1,12 @@
+import tkinter as tk
+import cv2
+import imutils
+import snooker_ball_tracker.settings as s
+import numpy as np
+import time
+from snooker_ball_tracker.views.application import Application
+
+
+if __name__ == "__main__":
+    app = Application()
+    app.mainloop()
